Fixed to compile with newer libs
Added a type signature to lazyElems. The inferred type signature uses
the FlexibleContexts language extension, and GHC >= 7.10.1 requires
said extension to be explicitly enabled, or the type signature to be
specified. The latter makes more sense to me.

Added an import for Data.Array.Unsafe. In array >= 5.0.0, unsafeThaw is
no longer exported by Data.Array.MArray, and is instead found in
Data.Array.Unsafe.